Read version info to render from package.json.
We already have a version information declared in package.json for NPM
description. Use it to render in web pages instead of hard coded string
or plain git commit hash.
Later on, the version information might get bumped when building
releases and reused to set tags in git and to Docker images.